requirements.txt
# Microsoft Agent Framework
agent-framework==1.0.0b251120
agent-framework-declarative==1.0.0b251120
# Requirements File
a2a-sdk==0.3.17
ag-ui-protocol==0.1.10
fastapi==0.121.3
fastmcp==2.13.1
httpx==0.28.1
pydantic==2.12.4
python-dotenv==1.2.1
pytz==2025.2
uvicorn==0.38.0
openai==2.8.1